Resueltos algunos problemas: 1- En la decision matrix de la fase de agregación se hacía el cálculo del inverso de una valoración que podía ser nula. Hay que ver cómo afecta eso al SA. 2- No se guardaban las valoraciones enteras cuando se unificaba. 3- Las valoraciones enteras tenían un valor entero y a la hora de calcular la normalización siempre salía 0, por lo tanto la unificación se hacía mal.

+++ sinbad2.aggregationoperator.arithmetic/src/sinbad2/aggregationoperator/arithmetic/valuation/IntegerOperator.java
@@ -14,7 +14,7 @@ public class IntegerOperator {
 	
 	public static Valuation aggregate(List<Valuation> valuations) {
 		IntegerValuation result = null;
-		long value = 0;
+		double value = 0;
 		int size = valuations.size();
 		NumericIntegerDomain domain = null;
