+++ sinbad2.aggregationoperator.owa/src/sinbad2/aggregationoperator/owa/valuation/IntegerOperator.java
@@ -46,7 +46,7 @@ public class IntegerOperator {
 			}
 
 			result = (IntegerValuation) valuations.get(0).clone();
-			result.setValue((long) measure);
+			result.setValue(measure);
 		}
 		
 		return result;
